Move beacon state endpoints to a separate module.
