+++ beacon_node/http_api/src/beacon/mod.rs
@@ -1,0 +1,1 @@
+pub mod states;
