+++ beacon_node/http_api/src/utils.rs
@@ -1,0 +1,3 @@
+use warp::filters::BoxedFilter;
+
+pub type ResponseFilter = BoxedFilter<(warp::reply::Response,)>;
